Fix import paths after backend reorganization
- Update shared/models/index.js to use new model locations
- Fix Transaction model import paths for enums and Tag
- Fix ProjectBudget model import paths for enums and Tag
