+++ backend/src/project-budgets/models/ProjectBudget.js
@@ -1,7 +1,7 @@
 const mongoose = require('mongoose');
 const currencyExchangeService = require('../../foreign-currency/services/currencyExchangeService');
-const { TransactionType } = require('../../shared/constants/enums');
-const Tag = require('../../shared/models/Tag');
+const { TransactionType } = require('../../banking/constants/enums');
+const Tag = require('../../banking/models/Tag');
 const Transaction = require('../../banking/models/Transaction');
 
 const projectBudgetSchema = new mongoose.Schema({
