+++ backend/src/shared/models/index.js
@@ -1,10 +1,10 @@
-const Category = require('./Category');
-const SubCategory = require('./SubCategory');
-const ManualCategorized = require('./ManualCategorized');
+const Category = require('../../monthly-budgets/models/Category');
+const SubCategory = require('../../monthly-budgets/models/SubCategory');
+const ManualCategorized = require('../../banking/models/ManualCategorized');
 const Transaction = require('../../banking/models/Transaction');
 const User = require('../../auth/models/User');
 const BankAccount = require('../../banking/models/BankAccount');
-const Tag = require('./Tag');
+const Tag = require('../../banking/models/Tag');
 const CreditCard = require('../../banking/models/CreditCard');
 const MonthlyBudget = require('../../monthly-budgets/models/MonthlyBudget');
 const YearlyBudget = require('../../monthly-budgets/models/YearlyBudget');
